shared: introduce routing
Some of our customers wants to have routing logic with fallback.
They want to target nodes from a particular rack and if rack is not
available, switch to datacenter.
Or like-wise target datacenter and switch to whole cluster.

This commit introduces routing package `rt` that allows it to happen.
You define routing:
```
helper.NewHelper(
  ...
  helper.WithRoutingScope(rt.NewDCScope("wrongDC", rt.NewDCScope("datacenter1", nil)))
  ...
  )
```
And if automagically works when helper reads nodes list.